perf(optim): improve AdamW optimizer performance and precision
- Change AdamW parameters from double to float for better GPU performance
- Optimize spherical harmonics computation by splitting into separate kernel
- Replace double precision math with float equivalents where possible
- Add NVTX ranges for better profiling visibility
- Fix potential numerical stability issues in learning rate calculation
- Correct spherical harmonics reset indices

+++ tinygs/include/tinygs/optim/adamw.hpp
@@ -7,9 +7,9 @@ namespace tinygs {
 
 struct AdamWParameters {
   /// Shared parameters
-  double beta1 = 0.9;
-  double beta2 = 0.999;
-  double epsilon = 1e-8;
+  float beta1 = 0.9f;
+  float beta2 = 0.999f;
+  float epsilon = 1e-8f;
 
   /// AdaBound extension (from tiny-cuda-nn)
   bool enable_adabound = false;
